fix: make http_server example run on ubuntu
We used an Ubuntu VM on a Mac and found out that there we a few
deadlocking issues with the Tcp_connector, and the Reader/Writer
processes in the Socket.Flow.

This changeset removes the Handshake awaiting stage of the Protocol
module, and makes the reader/writer flow yield execution back to the
scheduler when yielding back to the reader/writer.

So for example, when the reader is _done reading_, the
callback that was registered with `flow.yield_writer` in the writer
process will be called. Before this, that callback simply sent a message
to the writer process to wake it up, but it failed to yield control back
to the scheduler.