feat: io-aware scheduling, selective receive, (#9)
In this commit we're moving away from `Unix.select` to use the `iomux`
package that implements a portable `poll/ppoll` interface for polling
file descriptors instead.

This allows us much more finer grained control over file descriptors
while being just as if not more performant.

We also make one big change: the `Socket` and `Io` modules implement
suspension via a new effect: `Syscall { name; mode; fd }` – this means
that a process can suspend when calling `Socket.recv` or `Socket.send`
instead of having to wait for an explicit message.

When we reach this effect, a Process is now marked as "Waiting for IO"
explicitly, denoting the socket, syscall, and mode. This helps when
waking up a process and distinguishing from a blocking `receive`
expression.

The reason to do this is that with the message passing approach, the gap
between `Socket_read` message being sent and being consumed is
arbitrarily large (from a new nanos to never, depending on a
scheduler's load and whether the process actually calls `receive` at
all), which means that by the time you receive it and try to act on it,
the socket may not be readable/writeable anymore.

The expected thing then would be to block the syscall, and suspend the
process. So instead we make that the default behavior and if we need to
we will introduce message-driven reads/writes in the future.

The current approach also allows us to immediately queue a process for
execution, and the immediate next action of that process would be to
read/write to that socket as expected, since that was the suspension
point.

I'm not entirely sure this is 100% the right approach, but it seems to
work fine.

Some other additions in this changeset are:

1. Rc – a reference counted value, currently not in use but definitely
   useful considering we're moving around values across threads

2. Dashmap – a poorly named atomic key-value list that I'm using as a
   poorman's concurrent hashmap. It's fast enough for what we're doing
   here, but most likely has bugs and won't work for other use-cases.

3. the `Socket` module is now part of Riot, but I kept the `Flow` module
   as part of the Http Server example. The intention is that we can
   build a web server outside of Riot that will implement a better
   Flow, but still keep the low level socket machinery closer to the
   core so we can eventually implement other strategies for polling
   (like dedicated IO schedulers).

4. a new IO module – this is doing the heavy lifting for the polling and
   figuring out what's what. Also wraps lower level calls to the `Unix`
   module.

5. Selective Receive through Refs:

Before message passing used a `select` function to mark messages as
`Take`, `Drop`, or `Skip`. This was flexible but has the downside of
scanning the entire mailbox on every reentry to the Receive effect.

To avoid this and still allow for selective receive patterns, we
introduce an evelope around every message that contains a uid, and allow
the `receive` call to take `Ref.t` argument to match against.

When calling receive with a ref, every message that has been created
_before_ this ref (and we would know based on the internal ref of the
message envelope) will be put on a wait-queue in the same order as it
arrived.

Effectively, you will jump to the first message that was send _after_
the ref was created.

Subsequent calls to receive without a ref will read first from the
wait-queue, so the message ordering is preserved.