fix: logger app blocks until ready
This change makes the Logger application startup process _block_ until
the logger is ready to take in requests. This means applications relying
on the logger will take a tiny bit longer to boot, but we can guarantee
that the log requests will not be dropped because of process spawn
order.

Close #82