feat: allow configurign top-level supervisor restart limits
This lets you start a runtime that will terminate if the top-level
application terminates, which is useful for CLIs that may crash and need
to stay crashed.

For long-running applications this helps configure how many times you
want the application to retry before fully shutting down.

Also add a small pretty printer to print out the config as debug
information when the runtime starts.